backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.28
pydantic>=2.7.0
python-dotenv
google-generativeai
